Move OpenCV3 provider to its own module
And add back the original provider in its original place
