+++ kivy/core/camera/camera_opencv.py
@@ -1,277 +1,103 @@
-"""
-OpenCV Camera: Implement CameraBase with OpenCV (cv2 module)
+'''
+OpenCV Camera: Implement CameraBase with OpenCV
+'''
 
-Author: Hugo Geoffroy "pistache" <h.geoffroy@eden-3d.org>
+#
+# TODO: make usage of thread or multiprocess
+#
 
-"""
-# Tasks
-# ----
-# - TODO: use threads or multiprocessing instead of rescheduling
+__all__ = ('CameraOpenCV')
 
-
-# Imports
-# -------
 from kivy.logger import Logger
 from kivy.clock import Clock
 from kivy.graphics.texture import Texture
 from kivy.core.camera import CameraBase
 
-from cv2 import (
-    VideoCapture,
-    CAP_PROP_FRAME_WIDTH as FRAME_WIDTH,
-    CAP_PROP_FRAME_HEIGHT as FRAME_HEIGHT,
-    CAP_PROP_FPS as FPS,
-)
+try:
+    import opencv as cv
+    import opencv.highgui as hg
+except ImportError:
+    import cv
 
-# Constants
-# ---------
-IMAGE_FORMAT = 'bgr'  # OpenCV image format
+    class Hg(object):
+        '''
+        On OSX, not only are the import names different, but the API also
+        differs. There is no module called 'highgui' but the names are directly
+        available in the 'cv' module. Some of them even have a different
+        names.
 
+        Therefore we use this proxy object.
+        '''
 
-# Exports
-# -------
-__all__ = ['IMAGE_FORMAT', 'CaptureError', 'CameraOpenCV', ]
+        def __getattr__(self, attr):
+            if attr.startswith('cv'):
+                attr = attr[2:]
+            got = getattr(cv, attr)
+            return got
 
-
-# Exception classes
-# -----------------
-class CaptureError(RuntimeError):
-    """Raised by the Camera methods & scheduled routines. Should be catched and
-    translated to a proper kivy logged exception, if it's possible to recover
-    and continue.
-
-    """
-    def __init__(self, message, camera):
-        """Initialize capture error exception. Store the camera as exception
-        attributes.
-
-        """
-        self.camera = camera
-
-        super().__init__(message)
+    hg = Hg()
 
 
-# Camera provider class
-# ---------------------
 class CameraOpenCV(CameraBase):
-    """Implementation of CameraBase using OpenCV
-
-    Uses the :mod:`cv2` module, and its :class:`cv2.VideoCapture` class.
-
-    """
+    '''Implementation of CameraBase using OpenCV
+    '''
 
     def __init__(self, **kwargs):
-        """Initialize OpenCV Camera provider"""
-        self.capture = VideoCapture()
-
-        if __debug__:
-            Logger.debug(
-                "Camera: initializing capture ({})'"
-                "".format(self.capture)
-            )
-
-        try:
-            super(CameraOpenCV, self).__init__(**kwargs)
-        except CaptureError as ex:
-            Logger.exception(
-                "Camera: Exception while initializing camera : {}"
-                "".format(ex)
-            )
+        self._device = None
+        super(CameraOpenCV, self).__init__(**kwargs)
 
     def init_camera(self):
-        """Acquire camera and initialize capture
+        # create the device
+        self._device = hg.cvCreateCameraCapture(self._index)
 
-        """
-        image_format = self._format = IMAGE_FORMAT
-        index = self._index
-        width, height = self._resolution
+        # Set preferred resolution
+        cv.SetCaptureProperty(self._device, cv.CV_CAP_PROP_FRAME_WIDTH,
+                              self.resolution[0])
+        cv.SetCaptureProperty(self._device, cv.CV_CAP_PROP_FRAME_HEIGHT,
+                              self.resolution[1])
 
-        if __debug__:
-            Logger.debug(
-                "Camera: initializing (index: {}, image format: {}), "
-                "asking for size {}x{}"
-                "".format(index, image_format, width, height)
-            )
+        # and get frame to check if it's ok
+        frame = hg.cvQueryFrame(self._device)
+        # Just set the resolution to the frame we just got, but don't use
+        # self.resolution for that as that would cause an infinite recursion
+        # with self.init_camera (but slowly as we'd have to always get a
+        # frame).
+        self._resolution = (int(frame.width), int(frame.height))
 
-        self.capture.open(index)
-
-        if __debug__:
-            Logger.debug("Camera: opened, setting resolution...")
-
-        self.capture.set(FRAME_WIDTH, width)
-        self.capture.set(FRAME_HEIGHT, height)
-
-        if __debug__:
-            Logger.debug("Camera: resolution set, grab & read...")
-
-        self.capture.grab()
-
-        ok, frame = self.capture.read()
-
-        if not ok:
-            raise CaptureError(
-                "Could not read image from camera", self.capture
-            )
-
-        frame_height = len(frame)
-        frame_width = len(frame[0])
-
-        if __debug__:
-            Logger.debug(
-                "Camera: first frame size: {}x{}"
-                "".format(frame_width, frame_height)
-            )
-
-        if width != frame_width or height != frame_height:
-            if __debug__:
-                Logger.debug(
-                    "Camera: size corrected by camera : {}x{}"
-                    "".format(frame_width, frame_height)
-                )
-            self._resolution = frame_width, frame_height
-
-        self.fps = self.capture.get(FPS)
-
-        # needed because FPS determines rescheduling rate
+        #get fps
+        self.fps = cv.GetCaptureProperty(self._device, cv.CV_CAP_PROP_FPS)
         if self.fps <= 0:
-            Logger.warning(
-                "invalid FPS ('{}') returned by camera, using 30 as default"
-                "".format(self.fps)
-            )
-            self.fps = 1 / 30.0
+            self.fps = 1 / 30.
 
         if not self.stopped:
-            if __debug__:
-                Logger.debug("starting camera (initializing)")
             self.start()
 
-    def _release(self):
-        """Release camera.
-
-        This method unschedules frame update routines, then closes the capture
-        the device and releases the camera.
-
-        """
-        if __debug__:
-            Logger.debug("Camera: releasing...")
-
-        if not self.stopped:
-            self.stop()
-
-        if __debug__:
-            Logger.debug("Camera: schedule stop of update tasks")
-
-        self.capture.release()
-
-        if __debug__:
-            Logger.debug("released camera")
-
-    def _update_frame(self, delta):
-        """Update GPU buffer with camera image data.
-
-        """
-        if __debug__:
-            Logger.debug(
-                "Camera: updating GPU buffer... (delta: {})".format(delta)
-            )
-
+    def _update(self, dt):
         if self.stopped:
-            # Don't update it camere stopped
-            if __debug__:
-                Logger.debug(
-                    "frame update skipped as camera is stopped"
-                )
             return
-
         if self._texture is None:
-            if __debug__:
-                Logger.debug(
-                    "Camera: creating initial texture"
-                )
-            # Create the initial texture
+            # Create the texture
             self._texture = Texture.create(self._resolution)
             self._texture.flip_vertical()
             self.dispatch('on_load')
         try:
-            # Read buffer
-            ok, frame = self.capture.read()
-
-            # ok: boolean indicating whether frame reading was successfull
-            # frame: numpy matrix of pixel color values
-
-            if not ok:
-                raise CaptureError("Could not read image from camera", self)
-
-        except Exception as ex:
-            Logger.exception(
-                "Error while reading frame from camera : {}"
-                .format(ex)
-            )
-        else:
-            self._buffer = frame.tostring()
-
-            if __debug__:
-                Logger.debug(
-                    "Camera: got new frame (size: {})"
-                    "".format(len(self._buffer))
-                )
-
-            # this copies the image data to the OpenGL buffer
-            #
-            # (using Texture._blit_buffer)
-            #
+            frame = hg.cvQueryFrame(self._device)
+            self._format = 'bgr'
+            try:
+                self._buffer = frame.imageData
+            except AttributeError:
+                # On OSX there is no imageData attribute but a tostring()
+                # method.
+                self._buffer = frame.tostring()
             self._copy_to_gpu()
-
-            if __debug__:
-                Logger.debug(
-                    "GPU buffer updated"
-                )
+        except:
+            Logger.exception('OpenCV: Couldn\'t get image from Camera')
 
     def start(self):
-        """Start frame updating.
-
-        This method is not blocking, the update routine is just scheduled in
-        Kivy's clock."""
-
-        if not self.capture.isOpened():
-            Logger.exception(
-                "Camera is not open, cannot schedule frame updates",
-            )
-            return
-
-        if __debug__:
-            Logger.debug("Camera: starting capture...")
-
         super(CameraOpenCV, self).start()
-
-        if __debug__:
-            Logger.debug(
-                "unschedule update & reschedule with current FPS ({})"
-                "".format(self.fps)
-            )
-
-        Clock.unschedule(self._update_frame)
-        Clock.schedule_interval(self._update_frame, 1 / self.fps)
-
-        if __debug__:
-            Logger.debug("Camera: capture started")
+        Clock.unschedule(self._update)
+        Clock.schedule_interval(self._update, self.fps)
 
     def stop(self):
-        """Stop frame updating. This does not release the camera.
-
-        This method is not blocking, the update routine is just unscheduled
-        from Kivy's clock.
-
-        """
-        if __debug__:
-            Logger.debug("Camera: stopping capture...")
-
         super(CameraOpenCV, self).stop()
-
-        if __debug__:
-            Logger.debug("Camera: unschedule next update")
-
-        Clock.unschedule(self._update_frame)
-
-        if __debug__:
-            Logger.debug("Camera: capture stopped")
+        Clock.unschedule(self._update)
